requirements.txt
# Hybrid Quantum-Driven TB Detection - Dependencies
# Install numpy first (required by torch); then the rest.
numpy>=1.24.0,<3
torch>=2.0.0
torchvision>=0.15.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
tqdm>=4.65.0
gradio>=4.0.0
pandas>=2.0.0
